+++ js/constant.js
@@ -1,0 +1,54 @@
+const POSTS_COUNT = 25;
+const AVATAR_MIN_COUNT = 1;
+const AVATAR_MAX_COUNT = 25;
+const LIKES_MIN_COUNT = 15;
+const LIKES_MAX_COUNT = 200;
+const COMMENT_MIN_COUNT = 0;
+const COMMENT_MAX_COUNT = 30;
+const COMMENT_MAX_ID = 500000;
+
+const COMENT_NAMES = [
+  'Анна',
+  'Александр',
+  'Светлана',
+  'Юра',
+  'Сергей',
+  'Полина',
+  'Олег',
+  'Андрей',
+  'Юля',
+  'Анастасия',
+  'Соня',
+  'Татьяна',
+  'Алексей',
+  'Артем',
+  'Вера',
+  'Ирина',
+  'Николай',
+  'Ева',
+  'Ольга',
+  'Алиса'
+];
+
+const POSTS_COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+  'Моя жизнь меняется, потому что меняю ее я.',
+  'Не слушай никого, лишь твое сердце, оно все знает.',
+  'Чтобы достичь новых берегов, мы должны плыть, а не дрейфовать.',
+  'Навстречу новым приключениям.',
+  'Каждое мгновенье жизни — еще одна возможность.',
+  'Пусть эта фотография будет здесь',
+  'Одна хорошая мысль утром меняет смысл целого дня.',
+  'Как мало нужно для счастья.',
+  'Этой фотографии описание не нужно'
+];
+
+export {POSTS_COUNT, AVATAR_MIN_COUNT, AVATAR_MAX_COUNT, LIKES_MIN_COUNT, LIKES_MAX_COUNT, COMMENT_MIN_COUNT, COMMENT_MAX_COUNT, COMMENT_MAX_ID, COMENT_NAMES, POSTS_COMMENTS, DESCRIPTIONS};
